:recycle: #32 ストック取得APIへのリクエストをcreatedからfetchに移動
