+++ app/types/node.d.ts
@@ -1,0 +1,6 @@
+declare namespace NodeJS {
+  interface Process {
+    server: boolean
+    browser: boolean
+  }
+}
